feat: Removed server side OrganizationProvider
