+++ components/console/src/app/(routes)/ledgers/ledgers-sheet.tsx
@@ -19,7 +19,7 @@ import { z } from 'zod'
 import { LoadingButton } from '@/components/ui/loading-button'
 import { useCreateLedger, useUpdateLedger } from '@/client/ledgers'
 import { LedgerDto } from '@/core/application/dto/ledger-dto'
-import { useOrganization } from '@/providers/organization-provider/organization-provider-client'
+import { useOrganization } from '@/providers/organization-provider'
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs'
 import { useToast } from '@/hooks/use-toast'
 import { getInitialValues } from '@/lib/form'
