+++ components/console/src/app/(routes)/ledgers/page.tsx
@@ -15,7 +15,7 @@ import { LedgersSheet } from './ledgers-sheet'
 import { useCreateUpdateSheet } from '@/components/sheet/use-create-update-sheet'
 import { useDeleteLedger, useListLedgers } from '@/client/ledgers'
 import { useConfirmDialog } from '@/components/confirmation-dialog/use-confirm-dialog'
-import { useOrganization } from '@/providers/organization-provider/organization-provider-client'
+import { useOrganization } from '@/providers/organization-provider'
 import { LedgersSkeleton } from './ledgers-skeleton'
 import { useQueryParams } from '@/hooks/use-query-params'
 import { getBreadcrumbPaths } from '@/components/breadcrumb/get-breadcrumb-paths'
