+++ components/console/src/app/(routes)/portfolios/portfolios-sheet.tsx
@@ -16,7 +16,7 @@ import { useCreatePortfolio, useUpdatePortfolio } from '@/client/portfolios'
 import { DialogProps } from '@radix-ui/react-dialog'
 import { PortfolioDto } from '@/core/application/dto/portfolio-dto'
 import { LoadingButton } from '@/components/ui/loading-button'
-import { useOrganization } from '@/providers/organization-provider/organization-provider-client'
+import { useOrganization } from '@/providers/organization-provider'
 import { MetadataField } from '@/components/form/metadata-field'
 import { InputField } from '@/components/form'
 import { portfolio } from '@/schema/portfolio'
