+++ components/console/src/app/(routes)/settings/organizations-tab-content.tsx
@@ -32,7 +32,7 @@ import ConfirmationDialog from '@/components/confirmation-dialog'
 import { EntityDataTable } from '@/components/entity-data-table'
 import { Skeleton } from '@/components/ui/skeleton'
 import { OrganizationResponseDto } from '@/core/application/dto/organization-dto'
-import { useOrganization } from '@/providers/organization-provider/organization-provider-client'
+import { useOrganization } from '@/providers/organization-provider'
 import { IdTableCell } from '@/components/table/id-table-cell'
 
 export const OrganizationsTabContent = () => {
