+++ components/console/src/app/(routes)/transactions/[transactionId]/transaction-data-tab.tsx
@@ -15,7 +15,7 @@ import { PageFooter, PageFooterSection } from '@/components/page-footer'
 import { Button } from '@/components/ui/button'
 import { LoadingButton } from '@/components/ui/loading-button'
 import { useUpdateTransaction } from '@/client/transactions'
-import { useOrganization } from '@/providers/organization-provider/organization-provider-client'
+import { useOrganization } from '@/providers/organization-provider'
 import { useToast } from '@/hooks/use-toast'
 import { useConfirmDialog } from '@/components/confirmation-dialog/use-confirm-dialog'
 import ConfirmationDialog from '@/components/confirmation-dialog'
