+++ components/console/src/app/(routes)/transactions/create/components/account-search-field.tsx
@@ -12,7 +12,7 @@ import {
   AutocompleteValue
 } from '@/components/ui/autocomplete'
 import { Paper } from '@/components/ui/paper'
-import { useOrganization } from '@/providers/organization-provider/organization-provider-client'
+import { useOrganization } from '@/providers/organization-provider'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { Plus } from 'lucide-react'
 import React, { useState } from 'react'
