+++ components/console/src/app/(routes)/transactions/create/transaction-review.tsx
@@ -21,7 +21,7 @@ import { useTransactionForm } from './transaction-form-provider'
 import { useIntl } from 'react-intl'
 import { Separator } from '@/components/ui/separator'
 import { useCreateTransaction } from '@/client/transactions'
-import { useOrganization } from '@/providers/organization-provider/organization-provider-client'
+import { useOrganization } from '@/providers/organization-provider'
 import { useToast } from '@/hooks/use-toast'
 import { useRouter } from 'next/navigation'
 import { useState } from 'react'
