+++ components/console/src/components/organization-switcher/index.tsx
@@ -6,7 +6,7 @@ import React, { useEffect } from 'react'
 import { useIntl } from 'react-intl'
 import { useSidebar } from '../sidebar/primitive'
 import { Skeleton } from '../ui/skeleton'
-import { useOrganization } from '@/providers/organization-provider/organization-provider-client'
+import { useOrganization } from '@/providers/organization-provider'
 import { OrganizationSwitcherContent } from './organization-switcher-content'
 import { SwitcherTrigger } from './organization-switcher-trigger'
 import LerianLogo from '@/svg/lerian-logo.svg'
